ockam-server/controller/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ockam-controller"
version = "0.1.0"
description = "Control-plane API for an Ockam server node (Mode A docker / Mode B host install)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "filelock>=3.13",
  # Lets pydantic resolve PEP 604 `X | None` syntax on Python 3.9
  "eval-type-backport>=0.2; python_version<'3.10'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "httpx>=0.27",
]

[project.scripts]
ockam-controller = "ockam_controller.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
